libvirt backend: randomize hostnames
Vagrant's behavior is to take <dirname>_default as the name for the libvirt
domain it's spawning, but that is always .sandstorm under vagrant-spk, which
leads to colliding domain names.

vagrant-libvirt provides this option to avoid that case, which adds a timestamp
and a random string to the domain, so there will be no more collisions.