Use NFS mounts with virtualbox and 9p mounts with libvirt
When `spk pack` is called, for each file in the sandstorm-files.list, spk will
open the file, mmap it, and close the file.  Ultimately, this results in a
memory mapping for every file that will eventually land in the archive.

If this is more than 4096 files, then the open() calls will start to fail with
ENFILE, since the vboxsf file table is 4096 entries.

This may require further adaptation to support Windows, which does not support
NFS (but does support network sharing via SMB).

Alternately, we can change `spk`'s access patterns to be less troublesome,
which would let us return to vboxsf for compatibility and minimal invasiveness.