Use debian base box
There is an effort in the Debian project to provide official, up-to-date
Vagrant base boxes:

https://wiki.debian.org/Teams/Cloud/VagrantBaseBoxes

Switching from Ubuntu to Debian is a good idea because:

1) util-linux in Jessie includes nsenter, which will allow for entering
   grain namespaces for easier inspection of the runtime environment.
2) The current official release uses systemd
3) It has a very small minimal installation footprint
4) The kernel is more up-to-date

Debian's default nginx fastcgi_params do not include SCRIPT_FILENAME,
so we need to additionally set that in our nginx config.

As Debian has completed the systemd migration, pidfiles now default to
being created under /run, which is guaranteed to be tmpfs.  We don't
need them to be written at all, so we disable the feature in php5-fpm.