Thoughts about how to enter a grain's runtime environment
devjoin will join the only extent sandstorm-http-bridge's namespaces

join.sh is smarter, and will figure out which apps are in dev mode, which
running grains belong to those apps, which sandstorm-supervisor process is
responsible for the grain in question, and then will join the namespace of that
supervisor's child, which can also speak native capnproto.  But there's not a
great way to embed all that in an ssh -c command.

Conflicts:
	vagrant-spk