Reduce MySQL disk usage
This brings disk usage down to ~15MB.  I don't think we'll be able to do much
better; InnoDB requires at least 10MB for its data file path, so short of using
a patched mysqld, we can't fix this.

We could consider maintaining a patched MySQL which gets automatically pulled
in for vagrant-spk users, if that seems like a worthwhile time investment.