Make meteor build the bundle on the guest fs
On Windows, there's a path limit that Meteor/npm frequently exceed.  When we
attempt to build a bundle under /opt/app/.sandstorm/, it fails because Windows
will not allow such a long path.

This patchset makes vagrant-spk build the app in the guest OS in the vagrant
user's homedir instead.