Install git by default in uwsgi stack
Pip packages can reference git remotes, which is actually done a decent bit in
the Python community.  This should help things be more likely to work out of
the box, and git doesn't pull in too many dependencies, so it shouldn't be too
painful for users that don't need it.

Closes #119