Move all vagrant passthrough subcommands under the "vm" subcommand
In an attempt to clean up the command surface of vagrant-spk, all commands
intended to forward directly to vagrant are now placed under the "vm"
subcommand.

The "up", "ssh", "destroy", "global-status", and "reload" commands are
deprecated in favor of the "vm <subcommand>" equivalents, but will continue to
work with a warning until some future release.

This change also means we can reasonably pass through *any* arguments passed to
`vagrant-spk vm` on to `vagrant`, so we're no longer in the business of
whitelisting arguments and reimplementing all the commands that `vagrant`
provides that people might want to use.