Fix executable path detection on Windows 10.
The pyinstaller bundle generates a program for which `sys.argv[0]` is not the
full path to the executable, but just the short name of the program.

In this case, we should use `sys.executable` to determine the path to the .exe
that is actually launching vagrant-spk, since that's where the stacks will be
located.