Invoke bash manually for build.sh and launcher.sh
Rationale: When I was personally attempting to experiment with a Sandstorm app
written by someone on Windows, specifically https://github.com/ThomasBassa/commafeed-sandstorm ,
I found that the build.sh was not marked executable in their git repository.

This works around that issue by always using bash to exexcute build.sh.

I also noticed I had the same problem with launcher.sh. This works around that
for the future, too.

I see this commit, or something like it, as required if we support
Windows for vagrant-spk, where in my opinion, the executable bit is so likely
to get dropped that we ought to not rely on it.