Fix Vagrantfile for 2.0.3+
With Vagrant 2.0.3 and up, the DEFAULT_SERVER_URL is frozen, so it cannot be changed. But earlier versions of Vagrant need this line to work. While we could mandate a minimum Vagrant version, this change would set the correct URL if it's wrong, and leave it alone if the Vagrant version is new enough not to need it.

In https://github.com/hashicorp/vagrant/issues/9442 this fix was posted by @reedy, credit where credit is due.