Fix LEMP stack, install MySQL
So Debian 9 installs MariaDB instead of MySQL. That breaks vagrant-spk, so for now, this adds mysql.com as an apt source and gets MySQL proper. MySQL expected an additional directory in var, and indicated that the old way of initalizing the database was deprecated, so we're going to try changing that too.

Additionally, this fixes PHP 7's sock file location as with the update-lesp patch.