Disable MySQL privilege system
Previous recommendations is to use the root account with no password, so I do not believe this presents an increased security risk. Those looking to use MySQL user authentication will need to customize the stack and can remove this when they set up user accounts, but for most cases, this should be adequate, pending Kenton's opinion.