Fix #251 and some bitrot in enter_grain
This is one commit instead of two because we didn't want to bloat the
repository with two updates to the executable. We should really just
remove the binary from the repo and build it; see #253

The bitrot:

With #![feature(alloc_system)] I get an error from the rust toolchain
that the feature has been removed. We don't appear to actually be using
this anyway, so let's just remove it.