Minor formatting tweak in Vagrantfile message
This is a really trivial thing, but: my terminal's link detection seems
to want to include the period after the url in the url itself, which
means just trying to right click & open the link does not work
correctly. This patch gets around this by formatting the text in a way
that doesn't have the period at the end. I believe it still scans ok.