Fix remote-fs.target dependency cycle
Adding the automount option fixes a dependency cycle that can cause
Sandstorm to fail to start on boot for some development VMs.

The host\x2ddot\x2dsandstorm.mount is wanted by the remote-fs.target,
however it needs the virtualbox-guest-utils.service to mount
successfully.  The virtualbox-guest-utils.service can be made to run
before remote-fs.target, but I have never managed to make it be ready
for mounting Virtualbox Synced Folders at boot time.  It complains of
shared memory errors.  All of this causes the remote-fs.target to fail,
which causes sandstorm.service to fail to start at boot time.

This solution enables the automount option.  /host-dot-sandstorm is
instead no longer required for completion of the remote-fs.target and is
mounted on demand for vagrant-spk.  Everyone is happy.